added test and delegated old test to same class
